Rename input.tsx til Input.tsx
